refactor: move L1Event to btcio crate cleanup unused
