+++ crates/btcio/src/lib.rs
@@ -2,6 +2,7 @@
 #![allow(dead_code)] // TODO: remove this once `get_height_blkid` and `deepest_block` are used.
 
 pub mod broadcaster;
+pub mod events;
 pub mod reader;
 pub mod rpc;
 pub mod status;
